# GO-Web-1/docs/swagger.yaml
definitions:
  handler.request:
    properties:
      activo:
        type: boolean
      altura:
        type: number
      apellido:
        type: string
      edad:
        type: integer
      email:
        type: string
      fechaCreacion:
        type: string
      nombre:
        type: string
    required:
    - activo
    - altura
    - apellido
    - edad
    - email
    - nombre
    type: object
  handler.requestUpdateDTO:
    properties:
      activo:
        type: boolean
      altura:
        type: number
      apellido:
        type: string
      edad:
        type: integer
      email:
        type: string
      fechaCreacion:
        type: string
      id:
        type: integer
      nombre:
        type: string
    required:
    - activo
    - altura
    - apellido
    - edad
    - email
    - id
    - nombre
    type: object
  handler.requestUpdateFieldsDTO:
    properties:
      apellido:
        type: string
      nombre:
        type: string
    required:
    - apellido
    - nombre
    type: object
  web.Response:
    properties:
      code:
        type: string
      data: {}
      error:
        type: string
    type: object
info:
  contact:
    name: Natali Martinez
    url: natali.martinez@mercadolibre.com
  description: This API handle MELI users
  termsOfService: https://developers.mercadolibre.com.ar/es_ar/terminos-y-condiciones
  title: MELI Bootcamp Go W6 - API
  version: "4.2"
paths:
  /users:
    get:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List users
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.Response'
      summary: Show a list of users
      tags:
      - Users
    post:
      consumes:
      - application/json
      parameters:
      - description: token requerido
        in: header
        name: token
        required: true
        type: string
      - description: User to Store
        in: body
        name: users
        required: true
        schema:
          $ref: '#/definitions/handler.request'
      produces:
      - application/json
      responses:
        "200":
          description: user create
          schema:
            $ref: '#/definitions/web.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.Response'
        "500":
          description: 'Internal server error '
          schema:
            $ref: '#/definitions/web.Response'
      summary: Store users
      tags:
      - Users
  /users/{id}:
    delete:
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: integer
      - description: Token
        in: header
        name: token
        required: true
        type: string
      responses:
        "200":
          description: user delete
          schema:
            $ref: '#/definitions/web.Response'
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.Response'
        "500":
          description: 'Internal server error '
          schema:
            $ref: '#/definitions/web.Response'
      summary: Delete user
      tags:
      - Users
    get:
      parameters:
      - description: Id user
        in: path
        name: id
        required: true
        type: integer
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: user details
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.Response'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/web.Response'
        "500":
          description: 'Internal server error '
          schema:
            $ref: '#/definitions/web.Response'
      summary: Show a user
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: This endpoint update field name and lastname from a user
      parameters:
      - description: Token header
        in: header
        name: token
        required: true
        type: string
      - description: User id
        in: path
        name: id
        required: true
        type: integer
      - description: User name and lastname
        in: body
        name: name&lastname
        required: true
        schema:
          $ref: '#/definitions/handler.requestUpdateFieldsDTO'
      produces:
      - application/json
      responses:
        "200":
          description: user update
          schema:
            $ref: '#/definitions/web.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.Response'
        "500":
          description: 'Internal server error '
          schema:
            $ref: '#/definitions/web.Response'
      summary: Update name lastname
      tags:
      - Users
    put:
      consumes:
      - application/json
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: integer
      - description: Token
        in: header
        name: token
        type: string
      - description: User to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.requestUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: user update
          schema:
            $ref: '#/definitions/web.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.Response'
        "500":
          description: 'Internal server error '
          schema:
            $ref: '#/definitions/web.Response'
      summary: Update user
      tags:
      - Users
swagger: "2.0"